neostats/requirements.txt
streamlit
sentence-transformers
faiss-cpu
numpy
requests
beautifulsoup4
